Add vitest
