+++ .github/workflows/vitest.yml
@@ -1,0 +1,22 @@
+name: Run Tests
+
+on: push
+    
+jobs:
+    test:
+        runs-on: ubuntu-latest
+
+        steps:
+            - name: Checkout code
+              uses: actions/checkout@v4
+
+            - name: Set up Node.js
+              uses: actions/setup-node@v4
+              with:
+                node-version: 20
+
+            - name: Install dependencies
+              run: npm ci
+              
+            - name: Run tests
+              run: npm run test
