+++ package.json
@@ -6,7 +6,8 @@
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "lint": "next lint"
+    "lint": "next lint",
+    "test": "vitest"
   },
   "dependencies": {
     "aws-amplify": "^6.6.3",
@@ -17,18 +18,23 @@
   "devDependencies": {
     "@aws-amplify/backend": "^1.3.1",
     "@aws-amplify/backend-cli": "^1.2.9",
+    "@testing-library/dom": "^10.4.0",
+    "@testing-library/react": "^16.0.1",
     "@types/node": "^20",
     "@types/react": "^18",
     "@types/react-dom": "^18",
+    "@vitejs/plugin-react": "^4.3.3",
     "aws-cdk": "^2.160.0",
     "aws-cdk-lib": "^2.160.0",
     "constructs": "^10.3.0",
     "esbuild": "^0.24.0",
     "eslint": "^8",
     "eslint-config-next": "14.2.14",
+    "jsdom": "^25.0.1",
     "postcss": "^8",
     "tailwindcss": "^3.4.1",
     "tsx": "^4.19.1",
-    "typescript": "^5.6.2"
+    "typescript": "^5.6.2",
+    "vitest": "^2.1.4"
   }
 }
